--- desktop/install_profile_other.go
@@ -1,7 +1,0 @@
-//go:build !windows
-
-package main
-
-func telemetryInstallProfile() string {
-	return metricBucket(detectInstallProfile().Mode)
-}
