--- desktop/startup_recovery_other.go
@@ -1,9 +1,0 @@
-//go:build !darwin
-
-package main
-
-import "reasonix/internal/repair"
-
-func preparePackagedStartupRecovery(_ *repair.StartupTracker, recommended, explicitSafeMode bool) (bool, bool) {
-	return explicitSafeMode || recommended, true
-}
