+++ desktop/updater_install.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 	"runtime"
 	"strings"
 
@@ -126,25 +125,6 @@ func selectUpdateAsset(m *update.Manifest, profile installProfile) (update.Asset
 	return update.Asset{}, "", false
 }
 
-// dirIsWritable reports whether the process can create a temporary file in dir.
-func dirIsWritable(dir string) bool {
-	if dir == "" {
-		return false
-	}
-	info, err := os.Stat(dir)
-	if err != nil || !info.IsDir() {
-		return false
-	}
-	f, err := os.CreateTemp(dir, ".reasonix-write-test-*")
-	if err != nil {
-		return false
-	}
-	name := f.Name()
-	_ = f.Close()
-	_ = os.Remove(name)
-	return true
-}
-
 func linuxDebHelperReady() bool {
 	if runtime.GOOS != "linux" {
 		return false
@@ -156,18 +136,6 @@ func linuxDebHelperReady() bool {
 		return false
 	}
 	return true
-}
-
-// resolveExecutablePath returns the real path of the running binary.
-func resolveExecutablePath() string {
-	exe, err := os.Executable()
-	if err != nil {
-		return ""
-	}
-	if resolved, err := filepath.EvalSymlinks(exe); err == nil {
-		exe = resolved
-	}
-	return exe
 }
 
 func manualDebInstallHint() string {
