+++ desktop/webkit_compat_other.go
@@ -2,8 +2,6 @@
 
 package main
 
-func configureWebKitRendererRecovery(bool) {}
-
 func scheduleWebKitSignalHandlerRepair() {}
 
 func repairWebKitSignalHandlers() {}
